src/backend/requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn
pydantic